Module gcore::exec

Sys calls related to the program execution flow.

Provides API for low-level async implementation.

Functions

block_height

Get the current block height.

block_timestamp

Get the current block timestamp.

exit

Terminate the execution of a program. The program and all corresponding data are removed from the storage. This is similiar to std::process::exit. value_destination specifies the address where all available to the program value should be transferred to. Maybe called in init method as well.

gas_available

Get the current value of the gas available for execution.

leave

Terminate the current message handling.

origin

Return the id of original user who initiated communication with blockchain, during which, currently processing message was created.

program_id

Return ID of the current program.

value_available

Get the total available value amount.

wait

Pause the current message handling.

wake

Resume previously paused message handling.