Function gcore::exec::gas_available

pub fn gas_available() -> u64

Get the current value of the gas available for execution.

Each message processing consumes gas, both on instructions execution and memory allocations. This function returns a value of the gas available for spending during current execution. Its use may help to avoid unexpected behaviors during the smart-contract execution in case of not enough gas available.

Examples

use gcore::exec;

// Perform work while `gas_available` is more than 1000
unsafe extern "C" fn handle() {
    while exec::gas_available() > 1000 {
        // ...
    }
}