gcore::msg
Function id

pub fn id() -> MessageId

Get an identifier of the message that is currently being processed.

One can get an identifier for the currently processing message; each send and reply function also returns a message identifier.

Examples

use gcore::{msg, MessageId};

#[unsafe(no_mangle)]
extern "C" fn handle() {
    let current_message_id = msg::id();
    if current_message_id != MessageId::zero() {
        msg::reply(b"Real message", 0).expect("Unable to reply");
    }
}