Function gcore::msg::reply

pub fn reply(payload: &[u8], gas_limit: u64, value: u128) -> MessageId

Send a new message as a reply to the message currently being processed.

Some programs can reply to other programs, i.e. check another program’s state and use it as a parameter for its own business logic MessageId. This function allows sending such replies, which are similar to standard messages in terms of payload and different only in the way the message processing is handled by a separate program function called handle_reply.

First argument is the reply message payload in bytes. Second argument is gas_limit

maximum gas allowed to be utilized during the reply message processing. Last argument value is the value to be transferred from the current program account to the reply message target account.

Reply message transactions will be posted only after processing is finished, similar to the standard message send.

Examples

use gcore::{exec, msg};

pub unsafe extern "C" fn handle() {
    // ...
    msg::reply(b"PING", exec::gas_available(), 0);
}

See also

reply_push function allows to form a reply message in parts.