Function gcore::msg::reply_code

pub fn reply_code() -> Result<ReplyCode>

Get the reply code of the message being processed.

This function is used in the reply handler to check whether the message was processed successfully or not.

Examples

use gcore::msg;

#[no_mangle]
extern "C" fn handle_reply() {
    let reply_code = msg::reply_code().expect("Unable to get reply code");
}