Function gcore::msg::send_push

pub fn send_push(handle: &MessageHandle, payload: &[u8])

Push a payload part of the message to be sent in parts.

Gear allows programs to work with messages in parts. This function adds a payload part to the message specified by message handle.

Examples

use gcore::{exec, msg};

pub unsafe extern "C" fn handle() {
    // ...
    let msg_handle = msg::send_init();
    msg::send_push(&msg_handle, b"PING");
    msg::send_commit(msg_handle, msg::source(), exec::gas_available(), 42);
}

See also

send allows to send a message in one step.

send_init, send_commit functions allows to form and send a message to send in parts.