gcore
Module prog

API for creating programs from Gear programs.

Once a Wasm code has been uploaded to the chain, the Gear program can create a new program from the code blob using its identifier.

To create a new program, you are to provide the code identifier obtained after running the upload_code extrinsic, unique salt (arbitrary data needed to instantiate several programs from one code blob), and the init message that consists at least from a payload and value.

Functions

create_program

Create a new program and returns its address.

create_program_delayed

Same as create_program, but creates a new program after the delay expressed in block count.

create_program_with_gas

Same as create_program, but with an explicit gas limit.

create_program_with_gas_delayed

Same as create_program_with_gas, but creates a new program after the delay expressed in block count.