Function gcore::prog::create_program_with_gas

pub fn create_program_with_gas(
    code_hash: CodeHash,
    salt: &[u8],
    payload: &[u8],
    gas_limit: u64,
    value: u128
) -> Result<ActorId>

Creates a new program and returns its address, with gas limit.

The function creates a program initialization message and, as any message send function in the crate, this one requires common additional data for message execution, such as:

payload that can be used in init function of the newly deployed “child” program; 2. gas_limit, provided for the program initialization;

value, sent with the message. Code of newly creating program must be represented as blake2b hash (code_hash parameter).

Examples

In order to generate an address for a new program salt must be provided. Control of salt uniqueness is fully on a program developer side.

Basically we can use “automatic” salt generation (“nonce”):

use gcore::{prog, CodeHash};

static mut NONCE: i32 = 0;

fn increase() {
    unsafe {
        NONCE += 1;
    }
}

fn get() -> i32 {
    unsafe { NONCE }
}

unsafe extern "C" fn handle() {
    let submitted_code: CodeHash =
        hex_literal::hex!("abf3746e72a6e8740bd9e12b879fbdd59e052cb390f116454e9116c22021ae4a")
            .into();
    let new_program_id =
        prog::create_program_with_gas(submitted_code, &get().to_le_bytes(), b"", 10_000, 0)
            .unwrap();
}

Another case for salt is to receive it as an input:

use gcore::{msg, prog};

unsafe extern "C" fn handle() {
    let mut salt = vec![0u8; msg::size()];
    msg::load(&mut salt[..]);
    let new_program_id = prog::create_program_with_gas(submitted_code, &salt, b"", 10_000, 0).unwrap();
}

What’s more, messages can be sent to a new program:

use gcore::{msg, prog};

unsafe extern "C" fn handle() {
    let new_program_id = prog::create_program_with_gas(submitted_code, &salt, b"", 10_000, 0).unwrap();
    msg::send_with_gas(new_program_id, b"payload for a new program", 10_000, 0).unwrap();
}