gcore

Module exec

Utility functions related to the current execution context or program execution flow.

This module also provides API for low-level async implementation.

Functions

block_height

Get the current block height.

block_timestamp

Get the current block timestamp.

env_vars

Get current version of environment variables.

exit

Terminate the execution of a program.

gas_available

Get the current amount of gas available for execution.

leave

Break the current execution.

program_id

Return the identifier of the current program.

random

Get the random seed, along with the block number from which it is determinable by chain observers.

reply_deposit

Provide gas deposit from current message to handle reply message on given message id.

reserve_gas

Reserve the amount of gas for further usage.

system_reserve_gas

Reserve the amount of gas for system usage.

unreserve_gas

Unreserve gas identified by ReservationId.

value_available

Get the total available value amount.

wait

Pause the current message handling.

wait_for

Same as wait, but delays handling for a specific number of blocks.

wait_up_to

Same as wait, but delays handling for the maximum number of blocks that can be paid for and doesn’t exceed the given duration.

wake

Resume previously paused message handling.

wake_delayed

Same as wake, but executes after the delay expressed in block count.