Module gcore::msg

Messaging API for Gear programs.

This module contains sys calls API for incoming message processing and synchronous message sending. Messages are the main interface for communications between actors (users and programs). Every Gear program contains code for processing an incoming message. During a message processing a program can send messages to other programs and users including reply to the initial message.

Functions

exit_code

Get the exit code of the message being processed.

id

Obtain an identifier of the message currently being processed.

read

Get a payload of the message currently being processed.

reply

Send a new message as a reply to the message currently being processed.

reply_commit

Finalize and send a current reply message. Some programs can reply on their messages to other programs, i.e. check another program’s state and use it as a parameter for its own business logic. Basic implementation is covered in reply function.

reply_commit_delayed

Same as reply_commit, but sends delayed.

reply_commit_with_gas

Same as reply_commit, but with explicit gas limit.

reply_commit_with_gas_delayed

Same as reply_commit_with_gas, but sends delayed.

reply_delayed

Same as reply, but sends delayed.

reply_push

Push a payload part to the current reply message.

reply_to

Get an identifier of the initial message which the current handle_reply function is called on.

reply_with_gas

Same as reply, but with explicit gas limit.

reply_with_gas_delayed

Same as reply_with_gas, but sends delayed.

send

Send a new message to the program or user.

send_commit

Finalize and send message formed in parts.

send_commit_delayed

Same as send_commit, but with explicit gas limit.

send_commit_with_gas

Same as send_commit, but with explicit gas limit.

send_commit_with_gas_delayed

Same as send_commit_with_gas, but with explicit gas limit.

send_delayed

Same as send, but sends delayed.

send_init

Initialize a message to send formed in parts.

send_push

Push a payload part of the message to be sent in parts.

send_with_gas

Same as send, but with explicit gas limit.

send_with_gas_delayed

Same as send_with_gas, but sends delayed.

size

Get the payload size of the message being processed.

source

Get the identifier of the message source (256-bit address).

value

Get the value associated with the message being processed.