Function gcore::msg::send

pub fn send(
    program: ActorId, 
    payload: &[u8], 
    gas_limit: u64, 
    value: u128
) -> MessageId

Send a new message to the program or user.

Gear allows programs to communicate to each other and users via messages. send function allows sending such messages.

First argument is the address of the target account. Second argument is message payload in bytes. Third argument is gas_limit - maximum gas allowed to be utilized during the sent message processing. Last argument is the value to be transferred from the current program account to the message target account. Send transaction will be posted only after the execution of processing is finished, similar to the reply message reply.

Examples

use gcore::{msg, ActorId};

pub unsafe extern "C" fn handle() {
    // ...
    let mut id: [u8; 32] = [0; 32];
    for i in 0..id.len() {
        id[i] = i as u8;
    }

    msg::send(ActorId(id), b"HELLO", 1000, 12345678);
}

See also

send_init,send_push, send_commit functions allows to form a message to send in parts.